training/requirements.txt
torch==2.1.2
transformers==4.38.0
peft==0.8.2
bitsandbytes==0.42.0
accelerate==0.26.1
datasets==2.17.0
google-cloud-storage==2.14.0
wandb==0.16.3
scipy==1.11.4
sentencepiece==0.1.99
protobuf==4.25.2
